src/main/java/com/akman/springbootdemo/rest_api/SpringBootDemoApplication.java
package com.akman.springbootdemo.rest_api;

import com.akman.springbootdemo.batch.config.BatchConfig;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
import lombok.RequiredArgsConstructor;
import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;
import org.springframework.boot.autoconfigure.domain.EntityScan;
import org.springframework.boot.builder.SpringApplicationBuilder;
import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
import org.springframework.transaction.annotation.EnableTransactionManagement;
import org.springframework.web.WebApplicationInitializer;

import javax.annotation.PostConstruct;

@EnableBatchProcessing
@SpringBootApplication(scanBasePackageClasses = {
        BatchConfig.class,
        PackageMarker.class,
        com.akman.springbootdemo.PackageMarker.class})
@EnableTransactionManagement
@EnableJpaRepositories(
        basePackages = {"com.akman.springbootdemo.repository"}
)
@EntityScan(basePackages = {"com.akman.springbootdemo.model"})
@RequiredArgsConstructor
public class SpringBootDemoApplication extends SpringBootServletInitializer implements WebApplicationInitializer {

    public static void main(String[] args) {
        SpringApplication.run(SpringBootDemoApplication.class, args);
    }

    @Override
    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
        return application.sources(SpringBootDemoApplication.class);
    }

    private final ObjectMapper objectMapper;

    @PostConstruct
    public void setUp() {
        objectMapper.registerModule(new JavaTimeModule());
    }
}